Added General.XX files
