+++ src/state.h
@@ -4,29 +4,9 @@
 
 #include "hardware.h"
 #include "configurations.h"
-#include "whitelist.h"
-#include "rfid_utility.h"
-#include "signalisation.h"
-#include "EdgeDetection.h"
 
 #ifndef ARDUINO_ACCESS_SYSTEM_STATE_
 #define ARDUINO_ACCESS_SYSTEM_STATE_
-
-struct badgePresent
-{
-    uint32_t uid;
-    bool isMaster;
-};
-
-namespace General
-{
-    extern badgePresent badge;
-    extern Whitelist whitelist;
-    extern Tag rfid;
-    extern Signalisation signalize;
-    extern uint8_t tagAvailableVal;
-    extern EdgeDetection tagAvailable;
-} // namespace General
 
 namespace State
 {
